+++ include/patricia.h
@@ -32,7 +32,7 @@
  * structure is where the mask and data are kept.
  */
 struct ptree_mask {
-	unsigned long pm_mask;
+	unsigned int pm_mask;
 	void *pm_data;
 };
 
@@ -41,7 +41,7 @@ struct ptree_mask {
  * Patricia tree node.
  */
 struct ptree {
-	unsigned long p_key;		/* Node key		*/
+	unsigned int p_key;		/* Node key		*/
 	struct ptree_mask *p_m;		/* Node masks		*/
 	unsigned char p_mlen;		/* Number of masks	*/
 	char p_b;			/* Bit to check		*/
@@ -52,6 +52,6 @@ struct ptree {
 
 extern struct ptree *pat_insert(struct ptree *n, struct ptree *head);
 extern int           pat_remove(struct ptree *n, struct ptree *head);
-extern struct ptree *pat_search(unsigned long key, struct ptree *head);
+extern struct ptree *pat_search(unsigned int key, struct ptree *head);
 
 #endif /* _PATRICIA_H_ */
